create `geom` lib for plotting elements
